Add MetaGraph,SigDef unit tests
